refactor: validator to validator package
