+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
+import ru.yandex.practicum.filmorate.validator.Validator;
 
 import javax.validation.Valid;
 import java.util.Collection;
